+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed tracker `finish()` hook to `onFinish()`
+
 ## [4.0.0-alpha.1] - 2026-07-31
 
 ### Changed